SD Card functions arranged; UART removed; code cleaning
SD Card functions have been separated.
UART functions have been removed.
Code cleaning.
