+++ STM32_HUMIDIFIER/STM32_HUMIDIFIER.ioc
@@ -34,7 +34,6 @@ Mcu.CPN=STM32F411RET6
 Mcu.Family=STM32F4
 Mcu.IP0=ADC1
 Mcu.IP1=DMA
-Mcu.IP10=USART2
 Mcu.IP2=FATFS
 Mcu.IP3=I2C1
 Mcu.IP4=NVIC
@@ -43,35 +42,33 @@ Mcu.IP6=SPI2
 Mcu.IP7=SPI3
 Mcu.IP8=SYS
 Mcu.IP9=TIM10
-Mcu.IPNb=11
+Mcu.IPNb=10
 Mcu.Name=STM32F411R(C-E)Tx
 Mcu.Package=LQFP64
 Mcu.Pin0=PH0 - OSC_IN
 Mcu.Pin1=PH1 - OSC_OUT
-Mcu.Pin10=PB10
-Mcu.Pin11=PA10
-Mcu.Pin12=PA13
-Mcu.Pin13=PA14
-Mcu.Pin14=PA15
-Mcu.Pin15=PC10
-Mcu.Pin16=PC11
-Mcu.Pin17=PC12
-Mcu.Pin18=PD2
-Mcu.Pin19=PB3
+Mcu.Pin10=PA13
+Mcu.Pin11=PA14
+Mcu.Pin12=PA15
+Mcu.Pin13=PC10
+Mcu.Pin14=PC11
+Mcu.Pin15=PC12
+Mcu.Pin16=PD2
+Mcu.Pin17=PB3
+Mcu.Pin18=PB8
+Mcu.Pin19=PB9
 Mcu.Pin2=PC2
-Mcu.Pin20=PB8
-Mcu.Pin21=PB9
-Mcu.Pin22=VP_FATFS_VS_Generic
-Mcu.Pin23=VP_SYS_VS_Systick
-Mcu.Pin24=VP_TIM10_VS_ClockSourceINT
+Mcu.Pin20=VP_FATFS_VS_Generic
+Mcu.Pin21=VP_SYS_VS_Systick
+Mcu.Pin22=VP_TIM10_VS_ClockSourceINT
 Mcu.Pin3=PC3
-Mcu.Pin4=PA2
-Mcu.Pin5=PA3
-Mcu.Pin6=PA6
-Mcu.Pin7=PA7
-Mcu.Pin8=PC4
-Mcu.Pin9=PB0
-Mcu.PinsNb=25
+Mcu.Pin4=PA6
+Mcu.Pin5=PA7
+Mcu.Pin6=PC4
+Mcu.Pin7=PB0
+Mcu.Pin8=PB10
+Mcu.Pin9=PA10
+Mcu.PinsNb=23
 Mcu.ThirdPartyNb=0
 Mcu.UserConstants=
 Mcu.UserName=STM32F411RETx
@@ -109,10 +106,6 @@ PA15.GPIOParameters=GPIO_Label
 PA15.GPIO_Label=OLED_RES
 PA15.Locked=true
 PA15.Signal=GPIO_Output
-PA2.Mode=Asynchronous
-PA2.Signal=USART2_TX
-PA3.Mode=Asynchronous
-PA3.Signal=USART2_RX
 PA6.Signal=ADCx_IN6
 PA7.Signal=ADCx_IN7
 PB0.GPIOParameters=GPIO_Label
@@ -193,7 +186,7 @@ ProjectManager.StackSize=0x400
 ProjectManager.TargetToolchain=STM32CubeIDE
 ProjectManager.ToolChainLocation=
 ProjectManager.UnderRoot=true
-ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-HAL-false,2-MX_GPIO_Init-GPIO-false-HAL-true,3-MX_DMA_Init-DMA-false-HAL-true,4-MX_SPI2_Init-SPI2-false-HAL-true,5-MX_FATFS_Init-FATFS-false-HAL-false,6-MX_I2C1_Init-I2C1-false-HAL-true,7-MX_SPI3_Init-SPI3-false-HAL-true,8-MX_TIM10_Init-TIM10-false-HAL-true,9-MX_ADC1_Init-ADC1-false-HAL-true
+ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-HAL-false,2-MX_GPIO_Init-GPIO-false-HAL-true,3-MX_DMA_Init-DMA-false-HAL-true,4-MX_SPI2_Init-SPI2-false-HAL-true,5-MX_FATFS_Init-FATFS-false-HAL-false,6-MX_I2C1_Init-I2C1-false-HAL-true,7-MX_SPI3_Init-SPI3-false-HAL-true,8-MX_TIM10_Init-TIM10-false-HAL-true,9-MX_ADC1_Init-ADC1-false-HAL-true,10-MX_USART2_UART_Init-USART2-false-HAL-true
 RCC.48MHZClocksFreq_Value=50000000
 RCC.AHBFreq_Value=100000000
 RCC.APB1CLKDivider=RCC_HCLK_DIV2
@@ -250,8 +243,6 @@ SPI3.VirtualType=VM_MASTER
 TIM10.IPParameters=Prescaler,Period
 TIM10.Period=9999
 TIM10.Prescaler=9999
-USART2.IPParameters=VirtualMode
-USART2.VirtualMode=VM_ASYNC
 VP_FATFS_VS_Generic.Mode=User_defined
 VP_FATFS_VS_Generic.Signal=FATFS_VS_Generic
 VP_SYS_VS_Systick.Mode=SysTick
